requirements.txt
requests
python-dotenv
langchain
openai
langchain_community
langchain-openai
streamlit
